+++ src/components/modules/Footer/Footer.js
@@ -1,0 +1,22 @@
+// Framework Imports
+import React from 'react'
+
+// CSS Imports
+import StyledFooter from './StyledFooter'
+
+/**
+ * This the Footer component.
+ */
+const Footer = () => {
+  return (
+    <StyledFooter>
+      <p>
+        <a href='https://www.karlchvojka.com'>
+          &copy; 2023 - Karl Chvojka
+        </a>
+      </p>
+    </StyledFooter>
+  )
+}
+
+export default Footer