+++ webpack.config.js
@@ -50,9 +50,8 @@ module.exports = {
       elements: path.resolve(__dirname, './src/components/elements'),
       globalAssets: path.resolve(__dirname, './src/globalAssets'),
       globalConfig: path.resolve(__dirname, './src/globalConfig'),
-      modules: path.resolve(__dirname, './src/components/modules'),
-      pages: path.resolve(__dirname, './src/components/pages'),
-      templates: path.resolve(__dirname, './src/components/templates'),
+      layouts: path.resolve(__dirname, './src/components/layouts'),
+      modules: path.resolve(__dirname, './src/components/modules')
     },
 
     extensions: ['.js', '.jsx', '.json'],