+++ CHANGELOG.md
@@ -106,5 +106,27 @@ Note: #3 was done first to establish folder/file settings for the mobile setup
 
 ### Added
 ### Changed
+./package.json
+  - Updated all packages
+  - Updated project info
+./README.md
+  - Added instructions for maintenence on the template
+
 #### Package updates:
 Webpack Upgrde v5.74.0 => v5.76.3
+Babel Upgrade v7.18.10 =>
+@babel/cli                       ^7.18.10  →  ^7.21.0
+@babel/core                      ^7.18.10  →  ^7.21.3
+@babel/eslint-parser              ^7.18.9  →  ^7.21.3
+@babel/plugin-transform-runtime  ^7.18.10  →  ^7.21.0
+@babel/runtime                    ^7.18.9  →  ^7.21.0
+babel-loader                       ^8.2.5  →   ^9.1.2
+eslint                            ^8.21.0  →  ^8.36.0
+eslint-config-prettier             ^8.5.0  →   ^8.8.0
+eslint-plugin-import              ^2.26.0  →  ^2.27.5
+eslint-plugin-jsx-a11y             ^6.6.1  →   ^6.7.1
+eslint-plugin-react               ^7.30.1  →  ^7.32.2
+html-loader                        ^4.1.0  →   ^4.2.0
+styled-components                  ^5.3.5  →   ^5.3.9
+webpack-cli                       ^4.10.0  →   ^5.0.1
+webpack-dev-server                 ^4.9.3  →  ^4.13.1