+++ CHANGELOG.md
@@ -6,26 +6,50 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.0] - 2022-08-03
-Initial building of the Template.
+## [0.1.0] - 2023-03-26
 
 ### Added
-- .gitignore
-- README.md 
+./src/App.test.js
+  - Implemented basic render test
+./src/components/layouts/SidebarLayout
+  - Added Layout with sidebar
+### Changed
+./package.json
+  - Updated all packages
+  - Updated project info
+./README.md
+  - Added instructions for maintenence on the template
+./src/globalConfig/GlobalStyles
+  - Added default font styles
 
-- Added basic install and running instructions to the Readme.
+#### Package updates:
+Webpack Upgrde v5.74.0 => v5.76.3
+Babel Upgrade v7.18.10 =>
+@babel/cli                       ^7.18.10  →  ^7.21.0
+@babel/core                      ^7.18.10  →  ^7.21.3
+@babel/eslint-parser              ^7.18.9  →  ^7.21.3
+@babel/plugin-transform-runtime  ^7.18.10  →  ^7.21.0
+@babel/runtime                    ^7.18.9  →  ^7.21.0
+babel-loader                       ^8.2.5  →   ^9.1.2
+eslint                            ^8.21.0  →  ^8.36.0
+eslint-config-prettier             ^8.5.0  →   ^8.8.0
+eslint-plugin-import              ^2.26.0  →  ^2.27.5
+eslint-plugin-jsx-a11y             ^6.6.1  →   ^6.7.1
+eslint-plugin-react               ^7.30.1  →  ^7.32.2
+html-loader                        ^4.1.0  →   ^4.2.0
+styled-components                  ^5.3.5  →   ^5.3.9
+webpack-cli                       ^4.10.0  →   ^5.0.1
+webpack-dev-server                 ^4.9.3  →  ^4.13.1
 
-## [1.0.1] - 2022-08-04
-### Added
-- Installed ESLINT
+#### Package Installs:
+@testing-library/jest-dom: ^5.16.5,
+@testing-library/react: ^14.0.0,
+babel-jest: ^29.5.0,
+jest: ^29.5.0,
+jest-environment-jsdom: ^29.5.0,
+react-test-renderer: ^18.2.0,
 
-## [1.0.2] - 2022-08-09
-### Added
-- Cypress (left on branch because cypress is broken)
-- Built Header component
-- Added Webpack and Eslint support for alias's.
-
-## [1.0.3] - 2022-08-29
+## [0.0.4] - 2022-08-29
 Issue fixing
 
 Note: #3 was done first to establish folder/file settings for the mobile setup
@@ -102,3 +126,20 @@ Note: #3 was done first to establish folder/file settings for the mobile setup
 #### #6
 - Removed /cypress folder
 
+## [0.0.3] - 2022-08-09
+### Added
+- Cypress (left on branch because cypress is broken)
+- Built Header component
+- Added Webpack and Eslint support for alias's.
+
+## [0.0.2] - 2022-08-04
+### Added
+- Installed ESLINT
+
+## [0.0.1] - 2022-08-03
+Initial building of the Template.
+
+### Added
+- .gitignore
+- README.md 
+- Added basic install and running instructions to the Readme.